rename seed folder seeding
