+++ test-setup/setup-memory-server.ts
@@ -1,6 +1,6 @@
 import { MongoClient } from "mongodb";
 import { db } from "../connection";
-import seed from "../src/seed/seed";
+import seed from "../src/seeding/seed";
 import { MongoMemoryServer } from "mongodb-memory-server";
 
 const users = require("../src/seed/data/users.json");
